requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
python-dotenv==1.1.1
pytz==2025.2
